Final Code
The user is now able to select movies to put into or take out of their nominations list. When the user makes 5 nominations and trys to makes more selections they are given an alert, and all the nomination buttons are disabled. When the user removes nominations then they are again allowed to make more selections.